Séparation en deux des composants Localisation
