+++ source/components/localisation/Localisation.tsx
@@ -5,41 +5,18 @@ import {
 	supportedRegion,
 } from 'Components/localisation/utils'
 import { Trans } from 'react-i18next'
-import { useDispatch, useSelector } from 'react-redux'
-import { resetLocalisation, setLocalisation } from '../../actions/actions'
+import { useDispatch } from 'react-redux'
+import { resetLocalisation } from '../../actions/actions'
 import { usePersistingState } from '../../components/utils/persistState'
-import { capitalise0 } from '../../utils'
-import IllustratedMessage from '../ui/IllustratedMessage'
-import NewTabSvg from '../utils/NewTabSvg'
+import RegionGrid from './RegionGrid'
 
 export default () => {
 	const [chosenIp, chooseIp] = usePersistingState('IP', undefined)
 	const localisation = useLocalisation(chosenIp)
 	const dispatch = useDispatch()
-	const currentLang = useSelector((state) => state.currentLang).toLowerCase()
-
-	const supportedRegions = useSelector((state) => state.supportedRegions)
 	const isSupported = supportedRegion(localisation?.country?.code)
 	const flag = getFlag(localisation?.country?.code)
 	const countryName = getCountryNameInCurrentLang(localisation)
-
-	// Regions displayed sorted alphabetically
-	const orderedSupportedRegions = Object.fromEntries(
-		Object.entries(supportedRegions)
-			// sort function from https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/sort
-			.sort((a, b) => {
-				const nameA = a[1][currentLang].nom.toUpperCase() // ignore upper and lowercase
-				const nameB = b[1][currentLang].nom.toUpperCase() // ignore upper and lowercase
-				if (nameA < nameB) {
-					return -1
-				}
-				if (nameA > nameB) {
-					return 1
-				}
-				// names must be equal
-				return 0
-			})
-	)
 
 	return (
 		<div>
@@ -115,61 +92,7 @@ export default () => {
 					</Trans>{' '}
 				</p>
 			)}
-			<details>
-				<summary>
-					<Trans>Choisir une autre région</Trans>
-				</summary>
-				<ul
-					css={`
-						columns: 3;
-						-webkit-columns: 3;
-						-moz-columns: 3;
-					`}
-				>
-					{Object.entries(orderedSupportedRegions).map(([code, params]) => (
-						<li
-							key={code}
-							onClick={() => {
-								const newLocalisation = {
-									country: { name: params[currentLang]?.nom, code },
-									userChosen: true,
-								}
-								dispatch(setLocalisation(newLocalisation))
-								dispatch({ type: 'SET_LOCALISATION_BANNERS_READ', regions: [] })
-							}}
-						>
-							<button
-								css={`
-									padding: 0;
-									font-size: 0.75rem;
-									color: var(--darkColor);
-								`}
-							>
-								{capitalise0(params[currentLang]?.nom)}
-							</button>
-						</li>
-					))}
-				</ul>
-				<IllustratedMessage
-					emoji="🌐"
-					message={
-						<div>
-							<p>
-								<Trans>
-									Envie de contribuer à une version pour votre région ?
-								</Trans>{' '}
-								<a
-									target="_blank"
-									href="https://github.com/datagir/nosgestesclimat/blob/master/INTERNATIONAL.md"
-								>
-									<Trans>Suivez le guide !</Trans>
-									<NewTabSvg />
-								</a>
-							</p>
-						</div>
-					}
-				/>
-			</details>
+			<RegionGrid />
 		</div>
 	)
 }
